+++ quicklib-mvvm/src/main/java/com/quicklib/android/mvvm/view/QuickActivity.kt
@@ -1,4 +1,4 @@
-package com.quicklib.android.mvvm
+package com.quicklib.android.mvvm.view
 
 import androidx.lifecycle.ViewModel
 import androidx.databinding.ViewDataBinding
@@ -7,8 +7,10 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import com.quicklib.android.mvvm.QuickView
+import com.quicklib.android.mvvm.viewmodel.QuickViewModel
 
-abstract class QuickActivity<VM : ViewModel, VDB : ViewDataBinding> : AppCompatActivity() {
+abstract class QuickActivity<VDB : ViewDataBinding, VM : ViewModel> : AppCompatActivity(), QuickView<VDB, VM> {
 
     lateinit var binding: VDB
     lateinit var viewModel: VM
@@ -22,6 +24,7 @@ abstract class QuickActivity<VM : ViewModel, VDB : ViewDataBinding> : AppCompatA
             bindingCreated?.let {
                 binding = it
                 binding.setLifecycleOwner(this)
+                onBindingReady(binding)
             } ?: run {
                 throw IllegalStateException("Unable to find the binding of your root view. You should use DataBindingUtil.setContentView() instead of regular setContentView()")
             }
@@ -38,6 +41,6 @@ abstract class QuickActivity<VM : ViewModel, VDB : ViewDataBinding> : AppCompatA
         }
     }
 
-    abstract fun getViewModelInstance(): VM
-    abstract fun onViewReady(savedInstanceState: Bundle?)
+    override fun onBindingReady(binding: VDB) {}
+    override fun onViewReady(savedInstanceState: Bundle?) {}
 }
