+++ quicklib-mvvm/src/main/java/com/quicklib/android/mvvm/viewmodel/QuickViewModel.kt
@@ -1,4 +1,4 @@
-package com.quicklib.android.mvvm
+package com.quicklib.android.mvvm.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.databinding.Observable
